Week 10

+++ Week 03/Practice PB/C_1_Powering_the_Hero_easy_version.cpp
@@ -12,7 +12,7 @@ void solve()
         cin >> v[i];
     }
     priority_queue<int> pq;
-    int sum = 0;
+    ll sum = 0;
     for(auto i:v)
     {
         if(i > 0)
